adding setup. closes iss/mp-ids#31
Currently it is not necessary to run `./setup.sh` as we do not seem to
require a virtual environment.

The `beemaster`-directory exists, because we discussed to have a common
namespace for the project. `connector` then represents the specific
submodule. It might be implemented to be executed directly.

The `beemaster/connector/__init__.py` file is a general import helper,
should the connector-module be imported in another project (e.g. above
mentioned script).