adding pre-commit; combining environments
As mentioned in [0][], client-side hooks are not shared. If we would
want to enforce them, we would need to change them on the server-side.
As of this, the `setup.sh` now symlinks the `pre-commit` correctly, so
that it will be triggered on a `git commit`. As described in [1][], you
can enforce strict behaviour with `git config --bool flake8.strict
true`.

The `tox.ini` may only be changed with extensive description of the
reasons behind the changes.

[0]: https://git-scm.com/book/en/v2/Customizing-Git-Git-Hooks
[1]: http://flake8.pycqa.org/en/latest/user/using-hooks.html