updating ConnConfig to convert default_config
As previously the DEFAULT_CONFIG was only present as a normal dict and
not a ConnConfig object, access on values, only defined by the defaul
config would have resulted into errors.

By redefining .update, the conversion routine has been extracted and can
be reused more easily. The DEFAULT_CONFIG will still be added only if
the ConnConfig object is the primary/root Object.