Add check for connection status
Before sending to the master, check the connection status.

Only implemented for the master fallback, as the
loadbalancing recognises not connected slaves and should therefore reconnected to
another slave or the master.

However, it could be discussed if this check makes sense for slaves too,
because of running conditions (like network outage).

iss/mp-ids#86